Started a Shutdown process
A power outage required us to shut everything down.  It turns out that even though there was documentation for it, the process did not go smoothly.   This the beginning of the documented procedure brought to script.  I also reliaze that there is going to be another branch for these "experiments"